@@ -1,5 +1,5 @@
 module github.com/dependabot/dependabot-core/go_modules/helpers
 
-go 1.18
+go 1.19
 
 require github.com/Masterminds/vcs v1.13.3
