@@ -1,3 +1,5 @@
-// +build ignore
+//go:build ignore
+
+// this file only includes new style 1.17 build tags.
 
 package vgotest
