@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
-require "bundler"
+gem "bundler", "~> 2.3"
+require "bundler/setup"
 require "json"
 
 $LOAD_PATH.unshift(File.expand_path("./lib", __dir__))
@@ -14,6 +15,7 @@ end
 # Bundler monkey patches
 require "definition_ruby_version_patch"
 require "definition_bundler_version_patch"
+require "endpoint_specification_patch"
 require "git_source_patch"
 
 require "functions"
