@@ -3,8 +3,8 @@
 module Dependabot
   module Bundler
     module Helpers
-      V1 = "1.17.3"
-      V2 = "2.3.18"
+      V1 = "1"
+      V2 = "2"
       # If we are updating a project with no Gemfile.lock, we default to the
       # newest version we support
       DEFAULT = V2
