@@ -734,7 +734,7 @@ RSpec.describe Dependabot::Bundler::FileParser do
       parser.parse
 
       expect(events.last.payload).to eq(
-        { ecosystem: "bundler", package_managers: { "bundler" => PackageManagerHelper.bundler_major_version } }
+        { ecosystem: "bundler", package_managers: { "bundler" => PackageManagerHelper.bundler_version } }
       )
     end
   end
