@@ -20,16 +20,12 @@ module PackageManagerHelper
   end
 
   def self.bundler_version
-    use_bundler_2? ? "2.3.18" : "1.17.3"
-  end
-
-  def self.bundler_major_version
-    bundler_version.split(".").first
+    use_bundler_2? ? "2" : "1"
   end
 end
 
 def bundler_project_dependency_files(project)
-  project_dependency_files(File.join("bundler#{PackageManagerHelper.bundler_major_version}", project))
+  project_dependency_files(File.join("bundler#{PackageManagerHelper.bundler_version}", project))
 end
 
 def bundler_project_dependency_file(project, filename:)
