@@ -3,7 +3,7 @@
 require "excon"
 require "dependabot/metadata_finders"
 require "dependabot/metadata_finders/base"
-require "dependabot/shared_helpers"
+require "dependabot/registry_client"
 
 module Dependabot
   module Cargo
@@ -50,12 +50,7 @@ module Dependabot
       def crates_listing
         return @crates_listing unless @crates_listing.nil?
 
-        response = Excon.get(
-          "https://crates.io/api/v1/crates/#{dependency.name}",
-          idempotent: true,
-          **SharedHelpers.excon_defaults
-        )
-
+        response = Dependabot::RegistryClient.get(url: "https://crates.io/api/v1/crates/#{dependency.name}")
         @crates_listing = JSON.parse(response.body)
       end
     end
