@@ -3,6 +3,7 @@
 require "excon"
 require "dependabot/cargo/update_checker"
 require "dependabot/update_checkers/version_filters"
+require "dependabot/registry_client"
 
 module Dependabot
   module Cargo
@@ -83,12 +84,7 @@ module Dependabot
         def crates_listing
           return @crates_listing unless @crates_listing.nil?
 
-          response = Excon.get(
-            "https://crates.io/api/v1/crates/#{dependency.name}",
-            idempotent: true,
-            **SharedHelpers.excon_defaults
-          )
-
+          response = Dependabot::RegistryClient.get(url: "https://crates.io/api/v1/crates/#{dependency.name}")
           @crates_listing = JSON.parse(response.body)
         end
 
