@@ -146,6 +146,7 @@ module Dependabot
 
         def build_details_tag(summary:, body:)
           # Azure DevOps does not support <details> tag (https://developercommunity.visualstudio.com/content/problem/608769/add-support-for-in-markdown.html)
+          # Bitbucket does not support <details> tag (https://jira.atlassian.com/browse/BCLOUD-20231)
           # CodeCommit does not support the <details> tag (no url available)
           if source_provider_supports_html?
             msg = "<details>\n<summary>#{summary}</summary>\n\n"
@@ -240,7 +241,7 @@ module Dependabot
         end
 
         def source_provider_supports_html?
-          !%w(azure codecommit).include?(source.provider)
+          !%w(azure bitbucket codecommit).include?(source.provider)
         end
 
         def sanitize_links_and_mentions(text, unsafe: false)
