@@ -28,7 +28,7 @@
 
 const Arborist = require('@npmcli/arborist')
 const nock = require('nock')
-const { inspect, promisify } = require('util');
+const { promisify } = require('util');
 const exec = promisify(require('child_process').exec)
 
 async function findVulnerableDependencies(directory, advisories) {
@@ -59,16 +59,25 @@ async function findVulnerableDependencies(directory, advisories) {
 
   try {
     const name = advisories[0].dependency_name
+    const response = {
+      dependency_name: name,
+      fix_updates: [],
+      top_level_ancestors: [],
+    }
     const auditReport = await arb.audit()
+    if (!auditReport.has(name)) {
+      if (auditReport.tree.children.has(name)) {
+        response.current_version = auditReport.tree.children.get(name).version
+      }
+      response.fix_available = false
+      return response
+    }
     const vuln = auditReport.get(name)
     const version = [...vuln.nodes][0].version
     const fixAvailable = vuln.fixAvailable
-    const response = {
-      dependency_name: name,
-      current_version: version,
-      fix_available: Boolean(fixAvailable),
-      fix_updates: [],
-    }
+
+    response.current_version = version
+    response.fix_available = Boolean(fixAvailable)
 
     if (!Boolean(fixAvailable)) {
       return response
@@ -76,23 +85,40 @@ async function findVulnerableDependencies(directory, advisories) {
 
     const chains = buildDependencyChains(auditReport, name)
 
-    // In order to consider the vuln dependency in question fixable,
+    // In order for the vuln dependency in question to be considered fixable,
     // all dependency chains originating from it must be fixable.
     if (chains.some((chain) => !Boolean(chain.fixAvailable))) {
       response.fix_available = false
       return response
     }
 
-    for (const chain of chains) {
-      const topMost = chain.items[0]
-      if (topMost.name === name) {
-        continue
-      }
+    const groupedFixUpdateChains = groupBy(chains, (chain) => chain.nodes[0].pkgid)
+    let topLevelAncestors = new Set()
+
+    for (const group of groupedFixUpdateChains.values()) {
+      const fixUpdateNode = group[0].nodes[0]
+      const groupTopLevelAncestors = group.reduce((anc, chain) => {
+        const topLevelNode = chain.nodes[chain.nodes.length - 1]
+        return anc.add(topLevelNode.name)
+      }, new Set())
+
+      // Add group's top-level ancestors to the set of all top-level ancestors of
+      // the vuln dependency in question.
+      topLevelAncestors = new Set([...topLevelAncestors, ...groupTopLevelAncestors])
+
+      // If a chain consists of only one node, chain.nodes[0].name == chain.nodes[chain.nodes.length-1].name.
+      // In such cases, don't include the fix update node as an ancestor of itself.
+      const fixUpdateNodeTopLevelAncestors =
+        [...groupTopLevelAncestors].filter((nodeName) => nodeName !== fixUpdateNode.name).sort()
+
       response.fix_updates.push({
-        dependency_name: topMost.name,
-        current_version: topMost.version,
+        dependency_name: fixUpdateNode.name,
+        current_version: fixUpdateNode.version,
+        top_level_ancestors: fixUpdateNodeTopLevelAncestors,
       })
     }
+
+    response.top_level_ancestors = [...topLevelAncestors].sort()
 
     const fixTree = await arb.audit({
       fix: true,
@@ -135,75 +161,60 @@ function convertAdvisoriesToRegistryBulkFormat(advisories) {
   }, {})
 }
 
-/* Traverses all effects originating from the named dependency in the
- * audit report and returns an array of dependency chains rooted in the named
- * dependency,
+/* Returns an array of dependency chains rooted in the named dependency,
  *   [
  *    {
  *      fixAvailable: true | false | object,
- *      items: [
- *        { name: 'foo', version: '1.0.0' },
+ *      nodes: [
+ *        ArboristNode {
+ *          name: 'foo',
+ *          version: '1.0.0',
+ *          ...
+ *        },
  *        ...
  *      ]
  *    },
  *    ...
  *   ]
  *
- * The first item in the chain is always the top-most dependency affected by
- * the vulnerable dependency in question. The `fixAvailable` field
- * applies to the first item in the chain (if that item is fixable, then
- * every item after it must be fixable, too).
+ * The first node in each chain is the innermost dependency affected by the vuln;
+ * the `fixAvailable` field applies to this dependency. The last node in each
+ * chain is always a top-level dependency.
  */
 function buildDependencyChains(auditReport, name) {
-  const helper = (name, chain, visited) => {
-    // The vuln for this dependency.
-    const vuln = auditReport.get(name)
-
-    // The current version of this dependency.
-    const version = [...vuln.nodes][0].version
-
-    // The item that will represent this dependency in this chain.
-    const item = { name, version }
-
-    // Array of effects, excluding cycles.
-    const effects = [...vuln.effects]
-
-    if (visited.has(name)) {
-      // We've already visited this dependency in this chain, so we've detected a cycle.
-      // We currently throw when this happens. Ultimately we want to gracefully handle
-      // cycles and still return the recommended fix updates.
-      const source = chain.items[chain.items.length-1]
-      const message = `Cycle detected while traversing effects from ` +
-        `${source.name}@${source.version}: ` +
-        inspect([name, ...visited], {
-          breakLength: Infinity,
-          depth: 1,
-          maxStringLength: 255,
-        })
-      throw new Error(message)
+  const helper = (node, chain, visited) => {
+    if (!node) {
+      return []
     }
-
-    if (!effects.length) {
-      // If the current vuln has no effects, we've reached the end of this chain.
-      return [{ fixAvailable: vuln.fixAvailable, items: [item, ...chain.items] }]
+    if (visited.has(node.name)) {
+      // We've already seen this node; end path.
+      return []
     }
-
-    return effects.reduce((chains, effect) => {
-      return chains.concat(
-        helper(effect.name, { items: [item, ...chain.items] }, new Set([name, ...visited])))
+    if (auditReport.has(node.name)) {
+      const vuln = auditReport.get(node.name)
+      return [{ fixAvailable: vuln.fixAvailable, nodes: [node, ...chain.nodes] }]
+    }
+    if (!node.edgesOut.size) {
+      // This is a leaf node that is unaffected by the vuln; end path.
+      return []
+    }
+    return [...node.edgesOut.values()].reduce((chains, { to }) => {
+      // Only prepend current node to chain/visited if it's not the project root.
+      const newChain = node.isProjectRoot ? chain : { nodes: [node, ...chain.nodes] }
+      const newVisited = node.isProjectRoot ? visited : new Set([node.name, ...visited])
+      return chains.concat(helper(to, newChain, newVisited))
     }, [])
   }
+  return helper(auditReport.tree, { nodes: [] }, new Set())
+}
 
-  const chains = helper(name, { items: [] }, new Set())
-  const seen = new Set()
-  return chains.filter(chain => {
-    const head = chain.items[0]
-    if (seen.has(head.name)) {
-      return false
-    }
-    seen.add(head.name)
-    return true
-  })
+function groupBy(elems, fn) {
+  const groups = new Map()
+  for (const [index, elem] of [...elems].entries()) {
+    const key = fn(elem, index, elems)
+    groups.set(key, (groups.get(key) || []).concat([elem]))
+  }
+  return groups
 }
 
 async function loadNpmConfig() {
