@@ -36,11 +36,16 @@ module Dependabot
         #     * :dependency_name [String] the name of the blocking dependency
         #     * :current_version [String] the current version of the blocking dependency
         #     * :target_version [String] the target version of the blocking dependency
+        #     * :top_level_ancestors [Array<String>] the names of top-level dependencies with a transitive
+        #       dependency on the blocking dependency
+        #   * :top_level_ancestors [Array<String>] the names of all top-level dependencies with a transitive
+        #     dependency on the dependency
         def audit(dependency:, security_advisories:)
           fix_unavailable = {
             "dependency_name" => dependency.name,
             "fix_available" => false,
-            "fix_updates" => []
+            "fix_updates" => [],
+            "top_level_ancestors" => []
           }
 
           SharedHelpers.in_a_temporary_directory do
@@ -68,7 +73,7 @@ module Dependabot
               function: "npm:vulnerabilityAuditor",
               args: [Dir.pwd, vuln_versions]
             )
-            return fix_unavailable unless valid_audit_result?(audit_result, security_advisories)
+            return fix_unavailable unless viable_audit_result?(audit_result, security_advisories)
 
             audit_result
           end
@@ -81,15 +86,21 @@ module Dependabot
 
         attr_reader :dependency_files, :credentials
 
-        def valid_audit_result?(audit_result, security_advisories)
-          # we only need to check results that indicate a fix is available
-          return true unless audit_result["fix_available"]
+        def viable_audit_result?(audit_result, security_advisories)
+          validation_result = validate_audit_result(audit_result, security_advisories)
+          return true if validation_result == :viable
 
-          return false if vulnerable_dependency_removed?(audit_result)
-          return false if dependency_still_vulnerable?(audit_result, security_advisories)
-          return false if downgrades_dependencies?(audit_result)
+          Dependabot.logger.info("VulnerabilityAuditor: audit result not viable: #{validation_result}")
+          false
+        end
 
-          true
+        def validate_audit_result(audit_result, security_advisories)
+          return :fix_unavailable unless audit_result["fix_available"]
+          return :vulnerable_dependency_removed if vulnerable_dependency_removed?(audit_result)
+          return :dependency_still_vulnerable if dependency_still_vulnerable?(audit_result, security_advisories)
+          return :downgrades_dependencies if downgrades_dependencies?(audit_result)
+
+          :viable
         end
 
         def vulnerable_dependency_removed?(audit_result)
