@@ -37,9 +37,9 @@ RSpec.describe Dependabot::NpmAndYarn::UpdateChecker::VulnerabilityAuditor do
     end
 
     context "when a fix is available" do
-      let(:dependency_files) { project_dependency_files("npm8/locked_transitive_dependency") }
+      let(:dependency_files) { project_dependency_files("npm8/transitive_dependency_locked_by_intermediate") }
 
-      it "returns a hash with the target version and transitive updates to make" do
+      it "returns a hash with the top-level ancestors, target version, and transitive updates to make" do
         security_advisories = [
           Dependabot::SecurityAdvisory.new(
             dependency_name: dependency.name,
@@ -56,10 +56,12 @@ RSpec.describe Dependabot::NpmAndYarn::UpdateChecker::VulnerabilityAuditor do
             "target_version" => "1.0.1",
             "fix_available" => true,
             "fix_updates" => [{
-              "dependency_name" => "@dependabot-fixtures/npm-parent-dependency",
-              "current_version" => "2.0.0",
-              "target_version" => "2.0.2"
-            }]
+              "dependency_name" => "@dependabot-fixtures/npm-intermediate-dependency",
+              "current_version" => "0.0.1",
+              "target_version" => "0.0.2",
+              "top_level_ancestors" => ["@dependabot-fixtures/npm-parent-dependency"]
+            }],
+            "top_level_ancestors" => ["@dependabot-fixtures/npm-parent-dependency"]
           )
       end
     end
@@ -77,6 +79,7 @@ RSpec.describe Dependabot::NpmAndYarn::UpdateChecker::VulnerabilityAuditor do
           )
         ]
 
+        expect(Dependabot.logger).to receive(:info).with(/audit result not viable: vulnerable_dependency_removed/i)
         expect(subject.audit(dependency: dependency, security_advisories: security_advisories)).
           to include("fix_available" => false)
       end
@@ -102,6 +105,7 @@ RSpec.describe Dependabot::NpmAndYarn::UpdateChecker::VulnerabilityAuditor do
             "target_version" => "1.0.0"
           })
 
+        expect(Dependabot.logger).to receive(:info).with(/audit result not viable: dependency_still_vulnerable/i)
         expect(subject.audit(dependency: dependency, security_advisories: security_advisories)).
           to include("fix_available" => false)
       end
@@ -130,10 +134,13 @@ RSpec.describe Dependabot::NpmAndYarn::UpdateChecker::VulnerabilityAuditor do
             "fix_updates" => [{
               "dependency_name" => "@dependabot-fixtures/npm-parent-dependency",
               "current_version" => "2.0.0",
-              "target_version" => "1.0.2"
-            }]
+              "target_version" => "1.0.2",
+              "top_level_ancestors" => []
+            }],
+            "top_level_ancestors" => ["@dependabot-fixtures/npm-parent-dependency"]
           })
 
+        expect(Dependabot.logger).to receive(:info).with(/audit result not viable: downgrades_dependencies/i)
         expect(subject.audit(dependency: dependency, security_advisories: security_advisories)).
           to include("fix_available" => false)
       end
@@ -164,10 +171,10 @@ RSpec.describe Dependabot::NpmAndYarn::UpdateChecker::VulnerabilityAuditor do
       end
     end
 
-    context "when the native helper detects a vuln effects cycle" do
+    context "when the audit report contains a vuln effects cycle" do
       let(:dependency_files) { project_dependency_files("npm8/transitive_dependency_effects_cycle") }
 
-      it "returns fix_available => false and logs the error raised by the helper" do
+      it "returns a hash with the target version and transitive updates to make" do
         security_advisories = [
           Dependabot::SecurityAdvisory.new(
             dependency_name: dependency.name,
@@ -177,9 +184,20 @@ RSpec.describe Dependabot::NpmAndYarn::UpdateChecker::VulnerabilityAuditor do
           )
         ]
 
-        expect(Dependabot.logger).to receive(:info).with(/cycle detected/i)
         expect(subject.audit(dependency: dependency, security_advisories: security_advisories)).
-          to include("fix_available" => false)
+          to include(
+            "dependency_name" => dependency.name,
+            "current_version" => "1.0.0",
+            "target_version" => "1.0.1",
+            "fix_available" => true,
+            "fix_updates" => [{
+              "dependency_name" => "@dependabot-fixtures/npm-parent-dependency-4",
+              "current_version" => "1.0.0",
+              "target_version" => "2.0.0",
+              "top_level_ancestors" => []
+            }],
+            "top_level_ancestors" => ["@dependabot-fixtures/npm-parent-dependency-4"]
+          )
       end
     end
   end
